+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Fixed #285: class instances are now also converted by `toJS`. Also members defined on prototypes which are enumerable are converted.
 * Map keys are now always coerced to strings. Fixes #308
 * `when`, `autorun` and `autorunAsync` now accept custom debug names (see #293, by @jamiewinder)
+* Fixed #286: autorun's no longer stop working if an action throws an exception
 
 # 2.2.2: